Changed some constants so that the local proportions look more correct. (Note:
Previously, the player's horizontal motion looked significantly slower
than the same numerical amount of vertical motion. This was backwards of
what it should look like based on the foreshortening (i.e. the illusion
that we are looking at a 3D scene from behind and above).

To change the dimension-ratio at the location of the player, I had to
either fit more vertical length in a smaller space, or less horizontal
length in a larger space. I ended up doing both, but more of the latter
than the former.

Previously, there were 20 seconds of path visible at a time. In order to
give the player a sense of making progress, I didn't want to increase
that by very much. Increasing it to 24 seemed tolerable. Increasing it
to 40 seemed like too much.

I also experimented with changing the cylindrical perspective, but that
wasn't helpful – I had already found a good perspective, and you
couldn't really use that to fix the dimension-ratio at the player
location without, uh, "stealing" correct-looking perspective from other
areas.

This result makes the path appear twice as big on the screen, which is
probably good because I want to make there be more stuff on the path
(e.g. getting stars on the path shouldn't be THAT hard).