Refactored how we store computed data about machine futures, to simplify how modules will fit into the MachineTypeTrait API
Previously, MachineTypeTrait didn't allow caching of computed results
about futures, and I was considering adding an argument to output_flows
(and machine_momentary_visuals) that provided a mutable reference to a
cache. This approach is more elegant – the new future() function returns
all computed information, allowing the caching to be handled on the
graph-algorithms side, so that Module's MachineTypeTrait impl can be
agnostic regarding caching.